CAUD-204 FileBasedEventHandlerConfiguration.FileRetention#setMinFreeSpaceRequired() should take a long as its parameter

+++ audit/forgerock-audit-core/src/main/java/org/forgerock/audit/events/handlers/FileBasedEventHandlerConfiguration.java
@@ -263,7 +263,7 @@ public class FileBasedEventHandlerConfiguration extends EventHandlerConfiguratio
          * policy is disabled.
          * @param minFreeSpaceRequired The minimum free space the system must contain.
          */
-        public void setMinFreeSpaceRequired(final int minFreeSpaceRequired) {
+        public void setMinFreeSpaceRequired(final long minFreeSpaceRequired) {
             this.minFreeSpaceRequired = minFreeSpaceRequired;
         }
     }
